support idgen offset
