+++ util/service/ServBase_test.go
@@ -16,7 +16,7 @@ func TestIt(t *testing.T) {
 	//skey = "beauty"
 	var sb ServBase
 	var err error
-	sb, err = NewServBaseV2(configEtcd{etcds, "/roc"}, "niubi/fuck", skey)
+	sb, err = NewServBaseV2(configEtcd{etcds, "/roc"}, "niubi/fuck", skey, "",0)
 
 	if err != nil {
 		t.Errorf("create err:%s", err)
